Extract decoding item values for ISO mDL
