+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+Release 3.8.0:
+ - Extract credential classes for Mobile Driving Licence according to ISO 18013-5 into separate library, see <https://github.com/a-sit-plus/mobile-driving-licence-credential>
+
 Release 3.7.1:
  - SIOPv2: Support encrypting response objects, if requested by verifiers
  - Refactor `VerifiableCredentialSdJwt` to implement draft 03 of SD-JWT for VC
